factor out archivist and resource base classes
